requirements.txt
spleeter
opencv-contrib-python
python-vlc
pyserial
ffmpeg